Adding frontend

+++ src/main/java/com/myschool/demo/model/Questions.java
@@ -1,5 +1,7 @@
 package com.myschool.demo.model;
 
+import com.fasterxml.jackson.annotation.JsonBackReference;
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Data;
 
 import javax.persistence.*;
@@ -13,11 +15,12 @@ public class Questions {
     @Id
     @Column
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column
     private String Question;
 
+    //@JsonBackReference
     @OneToMany(mappedBy = "question")
     private List<Answers> answers;
